# docs/api.yaml
...
openapi: 3.1.1
servers:
  - url: http://127.0.0.1:4011/v1
info:
  title: A-Novel Authentication API.
  description: |
    The A-Novel authentication API, for internal use by A-Novel services.
  contact:
    name: A-Novel
    email: support@agoradesecrivains.com
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  version: 0.0.0
security:
  - bearerAuth: []

# ======================================================================================================================
# Tags
# ======================================================================================================================
...
paths:
...
  /session:
    get:
      tags:
        - session
      summary: Check the status of a session.
      description: |
        Takes an empty request with authorization headers, and check the validity of those headers. If the headers can
        be used to access any protected resource, the session is considered valid and the decoded claims are returned
        as a success response. Otherwise, an error will be sent, explaining why the session is invalid.
      operationId: checkSession
      responses:
        "200":
          description: The session is valid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claims"
        "401":
          description: |
            The authorization header is correctly set, but verification failed. The token is properly parsed, but its
            signature might be tampered or expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
    put:
      security: []
      tags:
        - session
        - credentials
      summary: Create a new session.
      description: |
        Create a new session, using a set of credentials. The provided credentials will be used to validate the 
        identity of the caller. Once the credentials have been verified, a token is issued. The access rights it grants
        may depend on the profile of the user.
      operationId: createSession
      requestBody:
        $ref: "#/components/requestBodies/LoginForm"
      responses:
        "200":
          description: The session has been created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "403":
          description: The provided credentials match a user, but authentication failed due to incorrect information.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: The provided credentials do not match any user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /session/anon:
    put:
      security: []
      tags:
        - session
      summary: Create a new anonymous session.
      description: |
        Create a new anonymous session. An anonymous session is delivered without constraint, and grants basic access
        to apis with low protection.
      operationId: createAnonSession
      responses:
        "200":
          description: The session has been created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "418":
          description: This route is not expected to fail under normal circumstances.
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /session/refresh:
    patch:
      security: []
      tags:
        - session
      summary: Refresh access token.
      description: |
        Takes a refresh token, and use it to issue a new access token.
      operationId: refreshSession
      parameters:
        - $ref: "#/components/parameters/RefreshToken"
        - $ref: "#/components/parameters/AccessToken"
      responses:
        "200":
          description: The session is valid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "403":
          description: |
            The authorization header is correctly set, but verification failed. The token is properly parsed, but its
            signature might be tampered or expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "422":
          description: The provided access token is not a refresh token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnprocessableEntityError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /short-code/register:
    put:
      tags:
        - shortCode
      security:
        - bearerAuth:
            - "register:request"
      summary: Set a new short code for user registration.
      description: |
        To prevent spam in our user database, registration must be done through a link sent by e-mail, so we can
        ensure this address is valid.
        When a user registers, the short code it received must be sent along with the registration payload. The email
        of the payload MUST match the email the short code was sent to, and is used to retrieve the short code.

        NOTE that this request does not verify the availability of an email. You may check this beforehand, using
        the `Email Exists` endpoint.

        If multiple registration links are requested for the same email, only the last one will be valid.
      operationId: requestRegistration
      requestBody:
        $ref: "#/components/requestBodies/RequestRegistrationForm"
      responses:
        "204":
          description: The short code has been sent successfully.
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "418":
          description: This route is not expected to fail under normal circumstances.
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /short-code/update-email:
    put:
      tags:
        - shortCode
      security:
        - bearerAuth:
            - "email:update:request"
      summary: Set a new short code for user email change.
      description: |
        Create a new short code for updating the email of an user. This short code is sent to the new address.
        If the user clicks on it, it should take it to a page that will forward the short code back to this API.
        Once done, the email associated to the user will be updated automatically.

        This route requires to be called by an authenticated user. Anonymous sessions cannot trigger an email
        update request.

        NOTE that this request does not verify the availability of an email. You may check this beforehand, using
        the `Email Exists` endpoint.

        If multiple email update links are requested for the same email, only the last one will be valid.
      operationId: requestEmailUpdate
      requestBody:
        $ref: "#/components/requestBodies/RequestEmailUpdateForm"
      responses:
        "204":
          description: The short code has been sent successfully.
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /short-code/update-password:
    put:
      tags:
        - shortCode
      security:
        - bearerAuth:
            - "password:reset:request"
      summary: Set a new short code for user password change.
      description: |
        Create a new short code for updating the password of an user. This short code is sent to the new address.
        If the user clicks on it, it should take it to a page that will forward the short code back to this API.
        Once done, the password of the user is updated.

        This route does not require authentication (although it requires at least an anonymous session). This
        allow users who forgot their password to reset it.

        If multiple password update links are requested for the same email, only the last one will be valid.
      operationId: requestPasswordReset
      requestBody:
        $ref: "#/components/requestBodies/RequestPasswordResetForm"
      responses:
        "204":
          description: The short code has been sent successfully.
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Access denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404":
          description: The provided email does not match any user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
  /credentials:
    put:
      tags:
        - credentials
      security:
        - bearerAuth:
            - "register"
      summary: Create a new user.
      description: |
        Create a new user. The form must contain a short code, that was sent through a registration link at the user
        desired email.

        On success, a valid access token is returned, that can be used to access higher-privilege routes.
      operationId: register
      requestBody:
        $ref: "#/components/requestBodies/RegisterForm"
      responses:
        "200":
          description: The user has been created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: The short code is not valid, either wrong or it does not exist at all for this email.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "410":
          description: The provided email is already in use.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
...
  /credentials/password:
    patch:
      tags:
        - credentials
      security:
        - bearerAuth:
            - "password:update"
      summary: Update the password of an user.
      description: |
        Update the password of an user. This route requires the original password of the user, to double check
        the identity of the caller.
      operationId: updatePassword
      requestBody:
        $ref: "#/components/requestBodies/UpdatePasswordForm"
      responses:
        "204":
          description: The password has been updated successfully.
        "401":
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: The provided current password does not match the actual password of the user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        default:
          description: An unexpected error occurred while processing the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  # ======================================================= SCHEMAS ====================================================
  schemas:
...
    Claims:
      type: object
      required:
        - roles
      properties:
        userID:
          type: string
          format: uuid
          description: The unique identifier of the user. Can be null if the session is anonymous.
          example: 29f71c01-5ae1-4b01-b729-e17488538e15
        roles:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Role"
          description: The roles granted by the token.
          example: ["auth:anon"]
        refreshTokenID:
          type: string
          description: The unique identifier of the refresh token. Can be null if the session is anonymous.
          example: 29f71c01-5ae1-4b01-b729-e17488538e15
...
    UnauthorizedError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message.
          example: The provided access token is not valid.
...
    UnexpectedError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message.
          example: An unexpected error occurred while processing the request.
...
  requestBodies:
    LoginForm:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LoginForm"
...